# Source/Identity/SchoolSaas.Web.Identity/wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "EduLink identity API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/utilityscope/getallutilityscopes": {
      "get": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_GetAllUtilityScopes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UtilityScopeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/getutilityscopesbycriteria": {
      "get": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_GetUtilityScopesByCriteria",
        "requestBody": {
          "x-name": "criteria",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilityScopeSearchCriteria"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UtilityScopeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/getutilityscopebyid": {
      "get": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_GetUtilityScopeById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilityScopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/createutilityscope": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_CreateUtilityScope",
        "parameters": [
          {
            "name": "Title",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "Description",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/assignutilityscopetorole": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_AssignUtilityScopeToRole",
        "parameters": [
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "UtilityScopeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/getutilityscopepermissions": {
      "get": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_GetUtilityScopePermissions",
        "parameters": [
          {
            "name": "utilityScopeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/assignpermissionstoutilityscope": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_AssignPermissionsToUtilityScope",
        "parameters": [
          {
            "name": "UtilityScopeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "PermissionIds",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/unassignutilityscopefromrole": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_UnassignUtilityScopeFromRole",
        "parameters": [
          {
            "name": "RoleId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "UtilityScopeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/unassignpermissionsfromutilityscope": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_UnassignPermissionsFromUtilityScope",
        "parameters": [
          {
            "name": "UtilityScopeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "PermissionIds",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/updateutilityscope": {
      "post": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_UpdateUtiltityScope",
        "parameters": [
          {
            "name": "UtilityId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "NewTitle",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "NewDescription",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/utilityscope/{utilityScopeId}": {
      "delete": {
        "tags": [
          "UtilityScope"
        ],
        "operationId": "UtilityScope_DeleteUtilityScope",
        "parameters": [
          {
            "name": "utilityScopeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "PermissionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "UtilityScopeDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UtilityScopeSearchCriteria": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "permissionId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}